Rename to InsertionStrategy
