+++ paging/src/commonMain/kotlin/org/mobilenativefoundation/store/paging5/InsertionStrategy.kt
@@ -1,0 +1,6 @@
+package org.mobilenativefoundation.store.paging5
+
+enum class InsertionStrategy {
+    APPEND,
+    PREPEND
+}
